pyproject.toml
[project]
name = "federated-satellites"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flwr>=1.19.0",
    "gdown>=5.2.0",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "omegaconf>=2.3.0",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "ray>=2.47.1",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "wandb>=0.21.0",
]
